[PATCH] ext3: error handling robustness
Fix a couple of ext3 error handling routines to not assume that the
superblock has valid journal and buffer_head pointers.  These functions are
called during mount and unmount and that may not be true.

This should fix the oops which Zwane saw when mounting a corrupt filesystem.

BKrev: 3e6d936fOHb_44eh28KQmnlm_jf2XQ