[PATCH] inline some mm functions for MMUless targets
This patch is the last of the MMUless support changes to the mm system.
It inlines for MMUless targets the following functions:

	vmalloc_to_page()
	follow_page()
	remap_page_range()

All of these return NULL for the non-MMU case. Doing the inline here
minimizes changes to assorted mm/ files. This patch was originally
from Christoph Hellwig.

BKrev: 3e6e2140U9fCLYcMETg-3XY0kHLWvQ