[PATCH] kNFSd: Introduce CROSSMNT flag for knfsd
Now that we have working up-calls to userspace,
CROSSMNT makes sense.
If CROSSMNT is set for an export, and we too a
LOOKUP which crosses a mountpoint, we initiate an
upcall to find out if and how that filesystem
is exported.

BKrev: 3e728bf1AjKL23KHP61zVIv7P0t6Qg