[PATCH] Pass the load address into ELF_PLAT_INIT()
Patch from Anton Blanchard <anton@samba.org>

With ppc64 64bit dynamic executables we have to relocate the contents of the
function descriptor.  Passing in the load address to ELF_PLAT_INIT allows us
to do this.

The patch allows ppc64 to run 64-bit executables and is a no-op for other
architectures.

BKrev: 3e778bf1dhFJmtWmqxJLLLqNelHVlw