[PATCH] Tighten CONFIG_NUMA preconditions
Patch from Martin J. Bligh and Dave Hansen

People with ordinary PCs are accidentally turning on NUMA support, and people
with NUMA machines are finding the NUMA option mysteriously disappearing.
This patch sets the defaults to sane things for everyone, and only allows you
to turn on NUMA with both SMP and 64Gb support on (it's useful for the
distros on non-Summit boxes, but not on their UP kernels ;-)).

I've also moved it below the highmem options, as it logically depends on
them, so this makes more sense.  For those searching for NUMA support on
*real* NUMA machine, Dave has provided some guiding comments to show them
what they messed up (it's totally non-obvious).  Hopefully this will stop
people's recent unfortunate foot-wounds (I think UP machines were defaulting
to NUMA on ...  oops).

BKrev: 3e778d940taKuqVk2rj0gD7CvaTO0Q