[PATCH] Fix nfsd_symlink() failure path
Patch from Andreas Gruenbacher <agruen@suse.de>

In both 2.5 and 2.4, the fs/nfsd/vfs.c:nfsd_symlink() function calls down to
notify_change().  If notify_change fails for some reason, the error code is
not converted to an nfs no-the-wire error code as is should.  The attached
patches fix that (one for 2.4, the other for 2.5).

BKrev: 3e778d9drWi5N3vP-jcAHN-zqAgXHw