[PATCH] raw driver: rewrite i_mapping only on final close
The recent fix to the raw driver wasn't quite right: it rewrites the
character-special inode's i_mapping to point back to itself on each close.
So any other currently-open handles against /dev/raw/rawN get a nasty
surprise.

Change it to only rewrite i_mapping on the final close.

Also, change it so that it only redirects its i_mapping on the initial open.
This is not necessary, but is neater.

BKrev: 3e7aa79aAfQR8E_5Wu93KAlEe5AwUw