[PATCH] fix powerbook media bay
This patch fixes a couple of bugs and compile errors in the powerbook
media bay driver.  It was getting initialized after the IDE subsystem,
whereas it needs to be initialized before so that the IDE subsystem
can see the CD-ROM drive in the bay.

BKrev: 3e7d476bFKmJFS5H13Bn4XMfoAAbqQ