[PATCH] Remove obsolete NFSD syscall varients.
Three varients of the NFSD system call use __kernel_* types in defining
the arg, which is bad: these types should not be exposed to userspace.

Particularly as __kernel_dev_t might change soon.

Two of the varients, getfh and uidmap, are not implemented, getfh is
deprecated and uidmap never was and never will be implemented.

This patch removes (almost) all trace of these two varients, including
all the 32bit-compat system call translators for various 64bit
platforms.

The third varient that uses __kernel_* types, export, is not so easy to
deal with, so it will be kept for later.

BKrev: 3e7e7b9avD4kqedDptOGrXiXKESoBg