[PATCH] don't include swap.h in mm.h
swap.h is basically the header for MM internals instead of the
public API (mm_internal.h would have been a better name..).  Stop
including it in mm.h - this only needs moving one function that
should be in swap.h anyway to the right place and fixing up a bunch
of places using it.

BKrev: 3e7e8064_NIwftJqvHrbvochZeSIKQ