Merge http://nfsclient.bkbits.net/linux-2.5 into penguin.transmeta.com:/home/penguin/torvalds/repositories/kernel/linux
2003/03/27 19:32:30+01:00 trond.myklebust
Fix clnt.c to skip re-encoding an RPC call in the case when
we're writing over TCP and have done a partial send.

2003/03/27 19:03:00+01:00 trond.myklebust
Fix a memory corruption bug in NFSv4 client.

2003/03/27 18:59:16+01:00 trond.myklebust
Fix a typo in auth_gss.c. Clean out an unused variable.

2003/03/27 18:43:56+01:00 trond.myklebust
  [PATCH] Fix misleading EIO on NFS client

  The following patch by Joe Korty removes an over-zealous check in the NFS
  read code that causes pages to be incorrectly marked with PG_error and
  hence causes an EIO to be returned to userland.

  The test is incorrect as it ignores the fact that we may be caching a
  write that will extend the file on the server (and hence will create a
  hole in the region concerned.).

2003/03/27 18:42:16+01:00 trond.myklebust
Fix the RPC debugging code so that it doesn't Oops if a task has
a null 'p_proc' procedure pointer.

2003/03/27 18:38:58+01:00 trond.myklebust
Micro-optimization: rename rpc_lookup_path() as rpc_lookup_parent()
and drop the 'flags' argument (it was always set too LOOKUP_PARENT).

If the user supplies and empty path, return -ENOENT.

2003/03/27 18:34:55+01:00 trond.myklebust
Fix xprt.c so that it resends RPC requests immediately after a timeout.
Doing this ensures that we keep probing the connection to the server
rather than just waiting for the entire congestion window to time out.
The latter can be very expensive due to the exponential backoff rule.

BKrev: 3e834a08niZCDIGd8Bcd9MQTNvBUww