[PATCH] acpi compile fix
ACPI is performing a spin_lock() on a `void *'.  That's OK when spin_lock is
implemented via an inline function.  But when it is implemented via macros
it causes compile-time breakage.

So cast it to the right type. It really should be fixed not to use opaque
handles, though.

BKrev: 3e8e5791TPEYgCB6Dy8uHZnWWVEcAQ