Add support for mapping NFSv4 remote user/group names into local unix-style uid/gids.
Note that this makes use of the RPC client upcall mechanism
(rpc_pipefs) to notify a userland daemon that does the actual mapping.
The results are then cached in the kernel.
The userland daemon can be downloaded from the CITI NFSv4 page at

     http://www.citi.umich.edu/projects/nfsv4/

BKrev: 3e91ffc3cgUNVbRmpIZ69Z4g92Qncw