Merge bk://are.twiddle.net/axp-2.5 into penguin.transmeta.com:/home/penguin/torvalds/repositories/kernel/linux
2003/04/15 13:31:11-07:00 drepper
[PATCH] unwinding for vsyscall code

Now that the kernel provides code user programs are executing directly
(I mean the vsyscall code on x86) it is necessary to add unwind
information for that code as well.  The unwind information is used not
only in C++ code.

This patch adds a AT_SYSINFO_EH_FRAME ELF aux-table value that points to
the unwinding block description for the sysinfo frame, and makes sure
the AT_* value is passed to applications.  It defines the static data
for the unwind blocks (two, one for int80 and the other for sysenter),
and finally adds code to copy the data in place.

2003/04/15 12:41:03-07:00 torvalds
Fix typo (and logic bug that the typo hid) in bit value
testing.

Found by 'sparse', my source parser tool.

BKrev: 3e9c729f7s1oDMAOxjzDwYAeVt_ZTA