Merge home.transmeta.com:/home/torvalds/v2.5/akpm into home.transmeta.com:/home/torvalds/v2.5/linux
2003/04/20 14:45:10-07:00 torvalds
Update ensoniq driver to return whether the interrupt was for it
or not.

2003/04/20 14:14:34-07:00 torvalds
Interrupt handlers should return whether the interrupt
was for them or not, so that the irq subsystem can properly
handle screaming shared interrupts.

So change the irq handlers to return a "irqretval_t", which
is either IRQ_HANDLED or IRQ_NONE.

2003/04/20 14:12:21-07:00 torvalds
Fix-ups for i830 from Arjan

BKrev: 3ea31585Fg2Kf3D2VvOnZFPRbdW9Gg