Merge bk://bk.arm.linux.org.uk/linux-2.5-pcmcia into home.transmeta.com:/home/torvalds/v2.5/linux
2003/04/23 22:04:13+01:00 rmk
[PCMCIA] Make cb_release_cis_mem() local to cardbus.c

The cardbus CIS parsing code does not use the PCMCIA resource
subsystem, so there isn't any point in freeing its memory when
we remove PCMCIA memory resources.  We also free CIS resources
immediately prior to calling cb_free().  We might as well move
the function call into cb_free(), thereby making all references
to cb_release_cis_mem() local to cardbus.c

2003/04/23 21:18:16+01:00 rmk
[PCMCIA] Don't cache CIS bytes found to be invalid.

Several PCMCIA cards I have here do not work correctly over a
suspend/resume cycle; the PCMCIA code believes that the card has
been changed in the slot, and therefore performs a remove/insert
cycle.

This seems to be because the card returns more or less random data
when reading memory space, leading to the CIS cache mismatching
the card data.  This in turn is caused because we try to read CIS
data from both the attribute and memory spaces, and we add the result
to the CIS cache whether or not the returned data was valid.

We therefore convert the CIS cache to use a linked list, and provide
a way to remove cached data from that list.  We also replace the
"s->cis_used=0;" construct with a function "destroy_cis_cache(s)"
which clearly describes what we're doing.

2003/04/23 20:59:33+01:00 proski
[PCMCIA] Fix oops in validate_mem when CONFIG_PCMCIA_PROBE=n

If I compile a recent 2.5.x kernel without CONFIG_ISA defined, I get
an oops in validate_mem().  Stack trace contains 0x6b6b6b6 - a clear
sign that freed memory is being accessed.

It's the second validate_mem() in drivers/pcmcia/rsrc_mgr.c - the one
used when CONFIG_PCMCIA_PROBE is not defined.  It turns out the memory
is freed in do_mem_probe() when it's called from validate_mem().

The solution is to use the same trick as in the first validate_mem().
This problem is quite serious and it's not specific to the plx9052
driver. I see it with yenta_socket as well.

2003/04/23 20:55:14+01:00 hch
[PCMCIA] remove unused files

From Christoph Hellwig

There's no need to keep the stubs around.

2003/04/23 20:51:31+01:00 proski
[PCMCIA] Fix compilation of cardmgr

Patch from Pavel Roskin

ds.h should not be including linux/device.h when compiling userspace
code.

BKrev: 3ea71535Nxp77CxQ2hM8Wnt61BoD8A